// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Urgency = {
  NORMAL: 'NORMAL',
  MEDIUM: 'MEDIUM'
} as const

export type Urgency = (typeof Urgency)[keyof typeof Urgency]


export const RecurrenceUnit = {
  HOUR: 'HOUR',
  DAY: 'DAY'
} as const

export type RecurrenceUnit = (typeof RecurrenceUnit)[keyof typeof RecurrenceUnit]


export const OccurrenceStatus = {
  PENDING: 'PENDING',
  COMPLETED: 'COMPLETED',
  MISSED: 'MISSED'
} as const

export type OccurrenceStatus = (typeof OccurrenceStatus)[keyof typeof OccurrenceStatus]


export const DailyStatus = {
  SUCCESSFUL: 'SUCCESSFUL',
  MISSED: 'MISSED',
  NO_ACTIVITY: 'NO_ACTIVITY',
  OPTIONAL_ONLY: 'OPTIONAL_ONLY'
} as const

export type DailyStatus = (typeof DailyStatus)[keyof typeof DailyStatus]